Abstracted some code out of DetectScript and modified the actual call to the powershell script. Cleaned up no longer needed code.
