+++ tasks/synopsys-detect-task/ts/model/IBlackduckConfiguration.ts
@@ -6,6 +6,5 @@ export interface IBlackduckConfiguration {
     blackduckUsername: string | undefined
     blackduckPassword: string | undefined
 
-    useProxy: boolean
     proxyInfo: IProxyInfo | undefined
 }
