+++ vss-extension.json
@@ -1,7 +1,7 @@
 {
   "manifestVersion": 1,
   "id": "synopsys-detect",
-  "version": "3.0.14",
+  "version": "3.0.19",
   "name": "Synopsys Detect",
   "publisher": "synopsys-detect",
   "public": true,
